@@ -1,10 +1,15 @@
-from typing import Optional
+"""Represent an element on the page using its locator and a description."""
 
-from screenpy import Actor
-from playwright.sync_api import Locator
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
 
 from .abilities import BrowseTheWebSynchronously
 from .exceptions import TargetingError
+
+if TYPE_CHECKING:
+    from playwright.sync_api import Locator
+    from screenpy import Actor
 
 
 class Target:
@@ -23,41 +28,61 @@ class Target:
         Target().located_by("#enter-todo-field")
     """
 
+    locator: str | None
+    _description: str | None
+
     @staticmethod
-    def the(name: str) -> "Target":
+    def the(name: str) -> Target:
         """Provide a human-readable description of the target."""
         return Target(name)
 
-    def located_by(self, locator: str) -> "Target":
+    def located_by(self, locator: str) -> Target:
         """Provide the Playwright locator which describes the element."""
         self.locator = locator
         return self
 
     @property
-    def target_name(self):
-        return self._description if self._description is not None else self.locator
+    def target_name(self) -> str:
+        """Get the name of the Target.
+
+        Returns:
+            The text representation of this Target.
+        """
+        return self._description or self.locator or "None"
 
     @target_name.setter
-    def target_name(self, value):
+    def target_name(self, value: str) -> None:
+        """Set the target_name.
+
+        Args:
+            value: the new description to use.
+        """
         self._description = value
 
-    @target_name.deleter
-    def target_name(self):
-        del self._description
-
     def found_by(self, the_actor: Actor) -> Locator:
+        """Get the Playwright Locator described by this Target.
+
+        Args:
+            the_actor: the Actor who should find this Locator.
+
+        Returns:
+            The Locator which describes the element.
+        """
         browse_the_web = the_actor.ability_to(BrowseTheWebSynchronously)
         if browse_the_web.current_page is None:
-            raise TargetingError(
-                #                          v              deep              v
-                f"There is no active page! {the_actor} cannot find the {self}."
-            )
+            msg = f"There is no active page! {the_actor} cannot find the {self}."
+            raise TargetingError(msg)
 
         return browse_the_web.current_page.locator(self.locator)
 
     def __repr__(self) -> str:
+        """Get a human-readable representation of this Target.
+
+        Returns:
+            A string representing this Target.
+        """
         return self.target_name
 
-    def __init__(self, name: Optional[str] = None) -> None:
+    def __init__(self, name: str | None = None) -> None:
         self._description = name
         self.locator = None
