@@ -1,11 +1,19 @@
-from screenpy import Actor
+"""Ask about the number of an element."""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
 from screenpy.pacing import beat
 
-from ..target import Target
+if TYPE_CHECKING:
+    from screenpy import Actor
+
+    from ..target import Target
 
 
 class Number:
-    """Get how many of an element are on the page.
+    """Ask how many of an element are on the page.
 
     Examples::
 
@@ -13,16 +21,32 @@ class Number:
     """
 
     @staticmethod
-    def of(target: Target) -> "Number":
+    def of(target: Target) -> Number:
+        """Supply the Target to count.
+
+        Args:
+            target: the Target which describes the element to count.
+
+        Returns:
+            A new Number instance.
+        """
         return Number(target)
 
     def describe(self) -> str:
-        """Describe the Question."""
+        """Describe the Question.
+
+        Returns:
+            A description of this Question.
+        """
         return f"The number of {self.target}."
 
     @beat("{} examines the Number of the {target}.")
-    def answered_by(self, the_actor: Actor) -> str:
-        """Direct the Actor to read the Number from the target."""
+    def answered_by(self, the_actor: Actor) -> int:
+        """Direct the Actor to read the Number from the target.
+
+        Returns:
+            The number of elements the Actor found.
+        """
         return self.target.found_by(the_actor).count()
 
     def __init__(self, target: Target) -> None:
